backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
databases[postgresql]
pydantic
requests
python-dotenv
